### 2022-05-31 #4444 @editorialbot
[PRE REVIEW]: HostPhot: global and local photometry of galaxies hosting supernovae or other transients
**Submitting author:** @temuller (<a href="http://orcid.org/0000-0003-3939-7167">Tomás E. Müller-Bravo</a>)
**Repository:** https://github.com/temuller/hostphot
**Branch with paper.md** (empty if default branch): main
**Version:** v2.0.0
