### 2022-06-09 d1db444
update journals on references
