+++ paper/paper.bib
@@ -16,7 +16,7 @@
          {Pierfederici}, Francesco and {Pontzen}, Andrew and {Ptak}, Andy and
          {Refsdal}, Brian and {Servillat}, Mathieu and {Streicher}, Ole},
         title = "{Astropy: A community Python package for astronomy}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2013,
         month = oct,
@@ -76,7 +76,7 @@ archivePrefix = {arXiv},
          {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
          {Zabalza}, V. and {Astropy Contributors}},
         title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = sep,
@@ -103,7 +103,7 @@ archivePrefix = {arXiv},
 	{The Astroquery collaboration} and {a subset of the astropy collaboration}
 	},
     title = "{astroquery: An Astronomical Web-querying Package in Python}",
-  journal = {\aj},
+  journal = {The Astronomical Journal},
 archivePrefix = "arXiv",
    eprint = {1901.04520},
  primaryClass = "astro-ph.IM",
@@ -121,7 +121,7 @@ archivePrefix = "arXiv",
 @article{Brout2019,
[…]
