### 2022-06-13 #4444 @arfon
:wave: @temuller – thanks for your submission to JOSS. Could you take a look a [this list of potential reviewers](https://bit.ly/joss-reviewers) and identify a few people who would be good candidates to review this submission? Also, please don't use the `@` at the front of GitHub handles (to avoid @mentioning them here).
