### 2022-06-15 #4444 @arfon
:wave: @zpace and @hechth – would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission under consideration is a Python package: _[HostPhot: global and local photometry of galaxies hosting supernovae or other transients](https://github.com/openjournals/joss-reviews/issues/4444#)_

Many thanks in advance ✨
