+++ src/hostphot/interactive_aperture.py
@@ -31,20 +31,20 @@ def _choose_workdir(workdir):
 #----------------------------------------
 class InteractiveAperture:
     """Class to interactively set the aperture to calculate photometry.
-
+    
     Parameters
     ----------
     name: str
         Name of the object to find the path of the fits file.
-    filt: str, default, ``g``
+    filt: str, default, `g`
         Filter used to initiate the plot.
-    filters: str, default, ``None``
+    filters: str, default, `None`
         Filters to use to load the fits files. If `None` use all
         the filters of the given survey.
-    survey: str, default ``PS1``
+    survey: str, default `PS1`
         Survey to use for the zero-points and pixel scale.
-    masked: bool, default ``True``
-        If ``True``, uses masked images.
+    masked: bool, default `True`
+        If `True`, uses masked images.
     """
     def __init__(self, name, filt='g', survey='PS1', filters=None, masked=True):
         self.name = name
@@ -55,28 +55,28 @@ class InteractiveAperture:
         else:
             check_filters_validity(filters, survey)
             self.filters = filters
-
+        
         if masked:
             self.masked = 'masked_'
         else:
             self.masked = ''
-
+        
         base_file = os.path.join(f'{self.masked}{survey}_{filt}.fits')
         self.fits_file = os.path.join(__workdir__, name, base_file)
-
-        self.ellipse_parameters = ['x', 'y', 'width',
+        
+        self.ellipse_parameters = ['x', 'y', 'width', 
                                   'height', 'angle']
-
+        
         # this is where the photometry is saved
         self.flux_phot = {f:np.nan for f in self.filters}
         self.flux_phot.update({f'{f}_err':np.nan for f in self.filters})
         self.mag_phot = {f:np.nan for f in self.filters}
         self.mag_phot.update({f'{f}_err':np.nan for f in self.filters})
-
+        
         # initiate plots + widgets
         self._initiate_plots()
         self._initiate_widgets()
-
+ 
     # -------------- Plots ------------------
     # ------
     # Images
@@ -87,7 +87,7 @@ class InteractiveAperture:
                            cmap='gray',
                            vmin=m-s, vmax=m+s,
                            origin='lower')
-
+        
     def _change_image(self, filt):
         """Changes the plotted image."""
         base_file = os.path.join(f'{self.masked}{self.survey}_{filt}.fits')
@@ -97,7 +97,7 @@ class InteractiveAperture:
         self.header = img[0].header
         self.data = img[0].data
         self.data = self.data.astype(np.float64)
-
+        
         self.bkg = sep.Background(self.data)
         self.err = self.bkg.globalrms
 
@@ -127,10 +127,10 @@ class InteractiveAperture:
             self.e.update({slider.description:change.new})
 
         # removes previous ellipse
-        self.ax.clear()
+        self.ax.clear()  
         self.ax.add_patch(self.e)
         self.ax.set_title(self.title)
-        self.im = self.draw_image()
+        self.im = self._draw_image()
 
     def _get_init_eparams(self):
         """Gets initial ellipse parameters."""
@@ -142,7 +142,7 @@ class InteractiveAperture:
                  'height':{'value':max_dim/2, 'min':0, 'max':max_dim},
                  'angle':{'value':0, 'min':-90, 'max':90},
                 }
-
+        
     def _get_eparams(self):
         """Gets the ellipse parameters."""
         self.eparams = {}
@@ -154,36 +154,36 @@ class InteractiveAperture:
         ix, iy = event.xdata, event.ydata
         self.sliders['x'].value = ix
         self.sliders['y'].value = iy
-
+        
     # ---------------
     # Master Function
     def _initiate_plots(self):
[…]
