### 2022-06-20 f89ffc8
do not make pool_download async
