### 2022-06-21 #4444 @krachyon
Hi @arfon and @temuller,
I'll preface this by saying that I don't actually have any academic credentials beyond a Master's degree (for which I technically don't even have the final grade yet). The focus of my work has been on the technical details of doing PSF-astrometry on (mostly simulated) point sources. So I feel a bit unprepared to judge the "Substantial scholarly effort" part as I don't know much about the field of research this software is made for. I do have opinions about and hopefully a bit of experience in software development, so I may be able to contribute something in that regard.

Let me know if this is an obstacle, I think I would be able to complete the review within the next 2-4 Weeks

As I noticed a few things that are not covered by the review criteria but would imo be valuable to mention, is it necessary to strictly separate that from comments about topics covered by the guidelines?

And as a meta-comment to the editors, it would be nice to link to a few examples of good review interactions on the documentation page to give a sense of how the process is supposed to play out.
