### 2022-06-23 a971f1e
correct typos and examples
