### 2022-06-26 #4444 @arfon
> Let me know if this is an obstacle, I think I would be able to complete the review within the next 2-4 Weeks.

Thanks @krachyon – thanks for the background, we'd love to have you as a reviewer.

> So I feel a bit unprepared to judge the "Substantial scholarly effort" part as I don't know much about the field of research this software is made for. I do have opinions about and hopefully a bit of experience in software development, so I may be able to contribute something in that regard.

Feel free to discuss this in your feedback in the main review thread I'm about to open below.

> As I noticed a few things that are not covered by the review criteria but would imo be valuable to mention, is it necessary to strictly separate that from comments about topics covered by the guidelines?

Definitely. You're welcome to provide any/all feedback in the review.

> And as a meta-comment to the editors, it would be nice to link to a few examples of good review interactions on the documentation page to give a sense of how the process is supposed to play out.

Yes, this is a great idea. We do have an overview of the review process, but I note this is on the 'editor' docs page (we should probably include this on the reviewer docs too: https://joss.readthedocs.io/en/latest/editing.html#visualization-of-editorial-flow

Linking to some good review examples too would be a great idea!

In summary though, JOSS reviews are iterative and our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. We also encourage reviewers to comment and ask questions on this thread.
