### 2022-06-26 #4508 @editorialbot
[REVIEW]: HostPhot: global and local photometry of galaxies hosting supernovae or other transients
**Submitting author:** @temuller (<a href="http://orcid.org/0000-0003-3939-7167">Tomás E. Müller-Bravo</a>)
**Repository:** https://github.com/temuller/hostphot
**Branch with paper.md** (empty if default branch): main
**Version:** v2.1.0
