### 2022-06-26 temuller/hostphot#2 @krachyon
[JOSS] comments about paper, documentation
- Could you mention what the difference between local vs global photometry is and what that implies for your code? It feels like one of the central parts of the problem to differentiate a Supernova's flux from the host galaxy so a few remarks about what that entails would be a nice starting point for someone reading your code.

- in "HostPhot is fast at calculating photometry (up to a few seconds per object)" It sounds a bit like you're impyling that you implement the photometry algorithms yourself. I feel you could make the relationship between your code and the used libraries a bit clearer, i.e. saying something like "after cleaning the image, photometry is calculated with `photutils`".
