### 2022-06-26 #4508 @krachyon
Created issues for collecting my comments: https://github.com/temuller/hostphot/issues/3 https://github.com/temuller/hostphot/issues/2 https://github.com/temuller/hostphot/issues/1
