### 2022-06-27 temuller/hostphot#3 @temuller
Thanks for such detailed comments. I'll try to implement them in future versions of this and other codes.
