+++ src/hostphot/utils.py
@@ -11,6 +11,7 @@ from astropy import wcs
 from astropy.stats import sigma_clipped_stats
 from photutils.utils import calc_total_error
 
+
 def calc_sky_unc(image, exptime):
     """Calculates the uncertainty of the image from the
     sky standard deviation, sigma-clipped STD.
@@ -28,19 +29,19 @@ def calc_sky_unc(image, exptime):
         Estimated error of the image.
     """
 
-    avg, sky, sky_std = sigma_clipped_stats(image[(image!= 0)],
-                                                        sigma=3.0)
+    avg, sky, sky_std = sigma_clipped_stats(image[(image != 0)], sigma=3.0)
     error = calc_total_error(image, sky_std, exptime)
 
     return error
 
+
 def pixel2pixel(x1, y1, img_wcs1, img_wcs2):
-    """Convert the pixel coordinates from one image to another.
-    """
+    """Convert the pixel coordinates from one image to another."""
     coord1 = img_wcs1.pixel_to_world(x1, y1)
     x2, y2 = img_wcs2.world_to_pixel(coord1)
 
     return x2, y2
+
 
 def check_survey_validity(survey):
     """Check whether the given survey is whithin the valid
@@ -51,9 +52,11 @@ def check_survey_validity(survey):
     survey: str
         Survey name: ``PS1``, ``DES`` or ``SDSS``.
     """
-    valid_surveys = ['PS1', 'DES', 'SDSS']
-    assert survey in valid_surveys, (f"survey '{survey}' not"
-                                     f" in {valid_surveys}")
+    valid_surveys = ["PS1", "DES", "SDSS"]
+    assert survey in valid_surveys, (
+        f"survey '{survey}' not" f" in {valid_surveys}"
+    )
+
 
 def get_survey_filters(survey):
     """Gets all the valid filters for the given survey.
@@ -69,12 +72,11 @@ def get_survey_filters(survey):
         Filters for the given survey.
     """
     check_survey_validity(survey)
-    filters_dict = {'PS1':'grizy',
-                    'DES':'grizY',
-                    'SDSS':'ugriz'}
+    filters_dict = {"PS1": "grizy", "DES": "grizY", "SDSS": "ugriz"}
     filters = filters_dict[survey]
 
     return filters
+
 
 def survey_zp(survey):
     """Returns the zero-point for a given survey.
@@ -93,12 +95,11 @@ def survey_zp(survey):
         Zero-point.
     """
     check_survey_validity(survey)
-    zp_dict = {'PS1':25 ,  # + 2.5*np.log10(exptime)
-               'DES':30,
-               'SDSS':22.5}
+    zp_dict = {"PS1": 25, "DES": 30, "SDSS": 22.5}  # + 2.5*np.log10(exptime)
     zp = zp_dict[survey]
 
     return zp
+
 
 def get_image_gain(header, survey):
     """Returns the gain from an image's header.
@@ -119,14 +120,15 @@ def get_image_gain(header, survey):
         Gain value.
     """
     check_survey_validity(survey)
-    if survey=='PS1':
-        gain = header['HIERARCH CELL.GAIN']
-    elif survey=='DES':
-        gain = header['GAIN']
-    elif survey=='SDSS':
+    if survey == "PS1":
+        gain = header["HIERARCH CELL.GAIN"]
+    elif survey == "DES":
+        gain = header["GAIN"]
+    elif survey == "SDSS":
         gain = 0.0
 
     return gain
+
 
 def get_image_readnoise(header, survey):
     """Returns the read noise from an image's header.
@@ -148,37 +150,37 @@ def get_image_readnoise(header, survey):
         Read noise value.
     """
[…]
