+++ src/hostphot/dust.py
@@ -9,6 +9,7 @@ import extinction
 import hostphot
 from .utils import integrate_filter, get_survey_filters, extract_filters
 
+
 def _download_dustmaps():
     """Downloads the dust maps for extinction calculation if they are not found
     locally.
@@ -16,22 +17,26 @@ def _download_dustmaps():
     mapsdir = hostphot.__path__[0]
 
     # check if files already exist locally
-    dust_files = [os.path.join(mapsdir,
-                'sfddata-master',
-                f'SFD_dust_4096_{sky}gp.fits') for sky in ['n', 's']]
-    mask_files = [os.path.join(mapsdir,
-                'sfddata-master',
-                f'SFD_mask_4096_{sky}gp.fits') for sky in ['n', 's']]
+    dust_files = [
+        os.path.join(mapsdir, "sfddata-master", f"SFD_dust_4096_{sky}gp.fits")
+        for sky in ["n", "s"]
+    ]
+    mask_files = [
+        os.path.join(mapsdir, "sfddata-master", f"SFD_mask_4096_{sky}gp.fits")
+        for sky in ["n", "s"]
+    ]
     maps_files = dust_files + mask_files
     existing_files = [os.path.isfile(file) for file in mask_files]
 
-    if not all(existing_files)==True:
+    if not all(existing_files) == True:
         # download dust maps
-        sfdmaps_url = 'https://github.com/kbarbary/sfddata/archive/master.tar.gz'
+        sfdmaps_url = (
+            "https://github.com/kbarbary/sfddata/archive/master.tar.gz"
+        )
         response = requests.get(sfdmaps_url)
 
-        master_tar = 'master.tar.gz'
-        with open(master_tar, 'wb') as file:
+        master_tar = "master.tar.gz"
+        with open(master_tar, "wb") as file:
             file.write(response.content)
 
         # extract tar file under mapsdir directory
@@ -40,8 +45,18 @@ def _download_dustmaps():
         tar.close()
         os.remove(master_tar)
 
-def deredden(wave, flux, ra, dec, scaling=0.86, reddening_law='fitzpatrick99',
-                                                    dustmaps_dir=None, r_v=3.1, ebv=None):
+
+def deredden(
+    wave,
+    flux,
+    ra,
+    dec,
+    scaling=0.86,
+    reddening_law="fitzpatrick99",
+    dustmaps_dir=None,
+    r_v=3.1,
+    ebv=None,
+):
     """Dereddens the given spectrum, given a right ascension and declination or :math:`E(B-V)`.
 
     Parameters
@@ -77,32 +92,36 @@ def deredden(wave, flux, ra, dec, scaling=0.86, reddening_law='fitzpatrick99',
     _download_dustmaps()
 
     hostphot_path = hostphot.__path__[0]
-    dustmaps_dir = os.path.join(hostphot_path, 'sfddata-master')
+    dustmaps_dir = os.path.join(hostphot_path, "sfddata-master")
 
     m = sfdmap.SFDMap(mapdir=dustmaps_dir, scaling=scaling)
-    ebv = m.ebv(ra, dec) # RA and DEC in degrees
-    a_v = r_v*ebv
+    ebv = m.ebv(ra, dec)  # RA and DEC in degrees
+    a_v = r_v * ebv
 
-    rl_list = ['ccm89', 'odonnell94', 'fitzpatrick99', 'calzetti00', 'fm07']
-    assert reddening_law in rl_list, f'Choose one of the available reddening laws: {rl_list}'
+    rl_list = ["ccm89", "odonnell94", "fitzpatrick99", "calzetti00", "fm07"]
+    assert (
+        reddening_law in rl_list
+    ), f"Choose one of the available reddening laws: {rl_list}"
 
[…]
