+++ tests/test_cutouts.py
@@ -1,18 +1,20 @@
 import unittest
 from hostphot.cutouts import download_images
 
-class TestHostPhot(unittest.TestCase):
 
+class TestHostPhot(unittest.TestCase):
     def test_cutouts(self):
-        sn_name = 'SN2004ey'
+        sn_name = "SN2004ey"
         ra = 327.28254
         dec = 0.44422
         size = 100
 
         # SDSS always fails in the tests
-        for survey in ['PS1', 'DES']:
-            download_images(sn_name, ra, dec, overwrite=True,
-                        size=size, survey=survey)
+        for survey in ["PS1", "DES"]:
+            download_images(
+                sn_name, ra, dec, overwrite=True, size=size, survey=survey
+            )
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()

[…]
