### 2022-06-27 temuller/hostphot#1 @krachyon
> Are you talking about files that are no longer needed but are saved in the history? Could you please rephrase this?

Yes, when looking at the git-history you can still see (and access) e.g. `images/SN2006bh/PS1_g.fits`, the file's contents are still automatically downloaded when the repository is cloned. If you happen to change them slightly, all versions that once existed have to be transferred on clone, this can blow up your repository size very quickly and is why typically you should not add binaries to version control.

If those files are something you did not want to publish, you'll have to delete them from the history. Some context and a solution to do that are here https://stackoverflow.com/questions/58161926/permanently-removing-a-file-from-git-history
