### 2022-06-28 temuller/hostphot#2 @temuller
Just to be clear, the images do not contain supernova flux as they were obtained mostly before, but sometimes after the supernova disappeared.
