### 2022-06-28 #4508 @temuller
Thanks again for the review. I have made the proper changes to the code+paper.
I am now using `black` for linting and added the recommended way of testing the code (with pytest). A section about how to contribute was added to the README as well,
The fits images from the git history were removed, as suggested.
About the warnings when running the tests, I don´t think that silencing them is a good idea as this could also silence future warnings that might pop up. These do not come from my end and I don't  know when (if at all) they are going to be "corrected".
I have also updated the paper following the comments of the referee. Let me know if you have further comments.
