### 2022-07-07 a0af868
cross-match optional + better plots
