+++ src/hostphot/image_masking.py
@@ -5,7 +5,8 @@ from matplotlib.patches import Ellipse
 
 import sep
 from astropy.io import fits
-from astropy import wcs
+from astropy import wcs, units as u
+from astropy.coordinates import SkyCoord, concatenate
 from astropy.convolution import (
     Gaussian2DKernel,
     convolve_fft,
@@ -20,7 +21,11 @@ from hostphot.objects_detect import (
     find_catalog_objects,
     cross_match,
 )
-from hostphot.utils import check_survey_validity, pixel2pixel
+from hostphot.utils import (check_survey_validity, pixel2pixel,
+                            update_axislabels)
+
+import warnings
+from astropy.utils.exceptions import AstropyWarning
 
 
 # ----------------------------------------
@@ -139,9 +144,11 @@ def create_mask(
     host_dec,
     filt,
     survey,
+    ra=None, dec=None,
     bkg_sub=False,
-    threshold=10,
-    sigma=20,
+    threshold=15,
+    sigma=8,
+    crossmatch=True,
     extract_params=False,
     common_params=None,
     save_plots=True,
@@ -153,20 +160,27 @@ def create_mask(
     name: str
         Name of the object to find the path of the fits file.
     host_ra: float
-        Host-galaxy Right ascension of the galaxy in degrees.
+        Host-galaxy right ascension in degrees.
     host_dec: float
-        Host-galaxy Declination of the galaxy in degrees.
+        Host-galaxy declination in degrees.
     filt: str
         Filter to use to load the fits file.
     survey: str
         Survey to use for the zero-points and correct filter path.
+    ra: float, default ``None´´
+       Right ascension of an object, in degrees. Used for plotting.
+    dec: float, default ``None´´
+       Declination of an object, in degrees. Used for plotting.
     bkg_sub: bool, default ``False``
         If ``True``, the image gets background subtracted.
-    threshold: float, default `10`
+    threshold: float, default `15`
         Threshold used by :func:`sep.extract()` to extract objects.
-    sigma: float, default ``20``
+    sigma: float, default ``8``
         Standard deviation in pixel units of the 2D Gaussian kernel
         used to convolve the image.
+    crossmatch: bool, default ``True``
+        If ``True``, the detected objects are cross-matched with a
+        Gaia catalog.
     extract_params: bool, default ``False``
         If ``True``, returns the parameters listed below.
     common_params: tuple, default ``None``
@@ -194,7 +208,9 @@ def create_mask(
 
     header = img[0].header
     data = img[0].data
-    img_wcs = wcs.WCS(header, naxis=2)
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore', AstropyWarning)
+        img_wcs = wcs.WCS(header, naxis=2)
 
     data = data.astype(np.float64)
     bkg = sep.Background(data)
@@ -210,9 +226,13 @@ def create_mask(
             data_sub, bkg_rms, host_ra, host_dec, threshold, img_wcs
         )
         # preprocessing: cross-match extracted objects with a catalog
-        # cat_coord = find_gaia_objects(host_ra, host_dec, img_wcs)
-        cat_coord = find_catalog_objects(host_ra, host_dec, img_wcs)
-        nogal_objs = cross_match(nogal_objs, img_wcs, cat_coord)
+        # using two Gaia catalogs as they do not always include the
+        # same objects.
+        if crossmatch:
+            cat_coord1 = find_gaia_objects(host_ra, host_dec, img_wcs)
+            cat_coord2 = find_catalog_objects(host_ra, host_dec, img_wcs)
[…]
