### 2022-07-08 a80de9a
changes in plotting and other minor things
