+++ src/hostphot/objects_detect.py
@@ -198,7 +198,7 @@ def plot_detected_objects(data, objects, scale, img_wcs, ra=None, dec=None, outf
 
     fig = plt.figure(figsize=(10, 10))
     ax = plt.subplot(projection=img_wcs)
-    overlay = update_axislabels(ax)
+    update_axislabels(ax)
 
     im = ax.imshow(
         data,
@@ -227,8 +227,8 @@ def plot_detected_objects(data, objects, scale, img_wcs, ra=None, dec=None, outf
         ax.scatter(px, py, marker='*', s=140, c='g')
 
     if outfile:
-        #plt.tight_layout()
-        plt.savefig(outfile)
+        plt.tight_layout()
+        plt.savefig(outfile, bbox_inches='tight')
         plt.close(fig)
     else:
         plt.show()
