### 2022-07-09 #4508 @krachyon
> Thanks again for the review. I have made the proper changes to the code+paper. I am now using `black` for linting and added the recommended way of testing the code (with pytest). A section about how to contribute was added to the README as well, The fits images from the git history were removed, as suggested.
👍

> About the warnings when running the tests, I don´t think that silencing them is a good idea as this could also silence future warnings that might pop up. These do not come from my end and I don't know when (if at all) they are going to be "corrected".

Not blaming you, I've also been annoyed by the amount of warnings you get from astropy for when you do anything with fits-io. It seems usually possible, if finicky,  to write specific enough filters so that you don't silence anything unexpected https://docs.python.org/3/library/warnings.html#warning-filter. My personal priority would be to keep the test output as clean as possible, but it's your call.
