### 2022-07-09 temuller/hostphot#1 @krachyon
In a recent commit you commented out the `__pycache__` entry in .gitignore, these dirs should not be added to version control. They contain machine-specific information like `&/home/tomas/hostphot/tests/__init__.py` which can cause pretty confusing errors for others.
