### 2022-07-11 #4508 @temuller
I have found that with astropy warnings (https://docs.astropy.org/en/stable/warnings.html), one can filter some of these annoying warnings coming from the header. However, as I said, this might not catch some future deprecation warnings. I'll try this for some time and see how it goes. This is already implemented in my developer branch, so I will proceed to merge it.
