+++ docs/examples/cutouts_example.rst
@@ -3,17 +3,20 @@
 Cutouts
 =======
 
-This module allows you to download image cutouts from :code:`PS1`, :code:`DES` and :code:`SDSS`. For this, the user can use :func:`download_images` and use the coordinates of an object:
+This module allows you to download image cutouts from :code:`PS1`, :code:`DES` and :code:`SDSS` (and maybe other surveys in the future). For this, the user can use :func:`download_images` and use the coordinates of an object:
 
 .. code:: python
 
 	from hostphot.cutouts import download_images
 
-	download_images(name='SN2004eo', ra=308.22579, dec=9.92853, survey='PS1')
+	name = 'SN2004eo'
+	host_ra, host_dec = 308.2092, 9.92755  # coords of host galaxy of SN2004eo
+	survey = 'PS1'
+	download_images(name, host_ra, host_dec, survey=survey)
 
-A new directory is created with the name ``name`` under the working directory (see :ref:`Working Directory <work_dir>`). The downloaded fits images will have the format ``<survey>_<filter>.fits``. If the filters are not specified, images in all the available filters (survey dependent) are downloaded.
+A new directory is created with the name ``name`` under the working directory (see :ref:`Working Directory <work_dir>`). The downloaded fits images will have the format ``<survey>_<filter>.fits``. If the filters are not specified, images in all the available filters (survey dependent) are downloaded. It is recommended to use the coordinates of the host galaxy to have it in the centre of the image. Sometimes, the transient can be very far from the host galaxy, so having separate images for the object and its host might be better.
 
-Let's check the downloaded image.
+Let's check the downloaded image:
 
 .. code:: python
 
