+++ docs/examples/preprocessing.rst
@@ -15,8 +15,7 @@ The user can perform image coadding in a single line:
 	from hostphot.coadd import coadd_images
 
 	coadd_filters = 'riz'
-	survey = 'PS1'
-	coadd_images('SN2004eo', coadd_filters, survey)
+	coadd_images(name, filters=coadd_filters, survey=survey)  # creates a new fits file
 
 This creates a new fits image under the object's directory, in this case, with the name ``PS1_riz.fits``. Coadding images is useful for common aperture photometry (see below). For the coadd, HostPhot makes use of `reproject <https://reproject.readthedocs.io/en/stable/index.html>`_.
 
@@ -31,18 +30,19 @@ Some low-redshift galaxy can have foreground stars "sitting" on top of them. Hos
 
 	from hostphot.image_masking import create_mask
 
-	host_ra, host_dec = 308.2092, 9.92755  # coods of host galaxy of SN2004eo
-	coadd_mask_params = create_mask(name, host_ra, host_dec, 
-                                         filt=coadd_filters, survey=survey, 
-                                         extract_params=True)
+	# one can extract the mask parameters from the coadd
+	# this also creates new fits files
+	coadd_mask_params = create_mask(name, host_ra, host_dec,
+                                	filt=coadd_filters, survey=survey,
+                                	extract_params=True)  
 
 	for filt in 'grizy':
-            create_mask(name, host_ra, host_dec, filt, survey=survey, 
-			common_params=coadd_mask_params)
+    		create_mask(name, host_ra, host_dec, filt, survey=survey,
+                	    common_params=coadd_mask_params)
 
 
 Note that the host-galaxy coordinates need to be provided so HostPhot knows which object not to mask out. The steps above create masked fits images, in this case, with the names ``masked_PS1_<filter>.fits``. See below an exmaple of this mask applied:
 
 .. image:: static/mask.png
 
-The sigma clipping step can be fine tuned by setting the parameter of ``threshold`` (how many sigmas above the background noise the sources are detected). Note that the images can be background subtracted by setting ``bkg_sub=True``, although this is not needed for the ``DES`` and ``PS1`` images downloaded by HostPhot.
+If the user is not happy with the result of the masking, there are a few parameters that can be adjusted. For instance, ``threshold`` sets the threshold (how many sigmas above the background noise) used by ``sep`` for detecting objects. Lowering it will allow the detection of fainter objects. ``sigma`` is the width of the gaussian used for convolving the image and masking the detected objects. If ``crossmatch`` is set to ``True``, the detected objects are cross-matched with the Gaia catalog and only those in common are kept. This is useful for very nearby host galaxies (e.g. that of SN 2011fe) so the structures of the galaxy are not maked out, artificially lowering its flux. Note that the images can be background subtracted by setting ``bkg_sub=True``, although this is not needed for the ``DES`` and ``PS1`` images downloaded by HostPhot.
