+++ src/hostphot/utils.py
@@ -313,6 +313,6 @@ def update_axislabels(ax):
             ax.coords[i].set_axislabel('RA (J2000)',
                                      fontsize=20)
         else:
-            ax.coords[i].set_axislabel('Dec (2000)',
+            ax.coords[i].set_axislabel('Dec (J2000)',
                                      fontsize=20)
             #ax.coords[i].set_ticklabel(rotation=65)
