+++ src/hostphot/dust.py
@@ -9,6 +9,8 @@ import extinction
 import hostphot
 from .utils import integrate_filter, get_survey_filters, extract_filters
 
+import warnings
+
 
 def _download_dustmaps():
     """Downloads the dust maps for extinction calculation if they are not found
@@ -26,7 +28,7 @@ def _download_dustmaps():
         for sky in ["n", "s"]
     ]
     maps_files = dust_files + mask_files
-    existing_files = [os.path.isfile(file) for file in mask_files]
+    existing_files = [os.path.isfile(file) for file in maps_files]
 
     if not all(existing_files) == True:
         # download dust maps
@@ -53,9 +55,7 @@ def deredden(
     dec,
     scaling=0.86,
     reddening_law="fitzpatrick99",
-    dustmaps_dir=None,
     r_v=3.1,
-    ebv=None,
 ):
     """Dereddens the given spectrum, given a right ascension and declination or :math:`E(B-V)`.
 
@@ -94,9 +94,11 @@ def deredden(
     hostphot_path = hostphot.__path__[0]
     dustmaps_dir = os.path.join(hostphot_path, "sfddata-master")
 
-    m = sfdmap.SFDMap(mapdir=dustmaps_dir, scaling=scaling)
-    ebv = m.ebv(ra, dec)  # RA and DEC in degrees
-    a_v = r_v * ebv
+    with warnings.catch_warnings():
+        warnings.filterwarnings("ignore", category=DeprecationWarning)
+        m = sfdmap.SFDMap(mapdir=dustmaps_dir, scaling=scaling)
+        ebv = m.ebv(ra, dec)  # RA and DEC in degrees
+        a_v = r_v * ebv
 
     rl_list = ["ccm89", "odonnell94", "fitzpatrick99", "calzetti00", "fm07"]
     assert (
@@ -159,13 +161,13 @@ def calc_extinction(
 
     # calculate extinction
     flux = 100
-    deredden_flux = deredden(
+    dereddened_flux = deredden(
         filter_wave, flux, ra, dec, scaling, reddening_law, r_v
     )
 
     f1 = integrate_filter(filter_wave, flux, filter_wave, filter_response)
     f2 = integrate_filter(
-        filter_wave, deredden_flux, filter_wave, filter_response
+        filter_wave, dereddened_flux, filter_wave, filter_response
     )
     A_ext = -2.5 * np.log10(f1 / f2)
 
