### 2022-07-12 #4508 @temuller
I managed to remove all the warnings. This are only ignored when reading the headers with WCS. Hopefully, this shouldn't break anything in future versions. All the changes have already been merged with the main branch.
