+++ docs/conf.py
@@ -52,6 +52,11 @@ pygments_style = "sphinx"
 # a list of builtin themes.
 #
 html_theme = "sphinx_rtd_theme"
+html_logo = "hostphot_logo.png"
+html_theme_options = {
+    'logo_only': False,
+    'display_version': True,
+}
 html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
 # Add any paths that contain custom static files (such as style sheets) here,
