### 2022-07-13 #4508 @krachyon
> I managed to remove all the warnings. This are only ignored when reading the headers with WCS. Hopefully, this shouldn't break anything in future versions. All the changes have already been merged with the main branch.

Cool, thanks for taking the extra time! With regards to the question marks in my checklist, maybe @Onoddil can have a closer look at those in their review.

As far as I'm concerned, everything looks good, so I would consider my review finished and in my opinion the software/paper should be accepted.
