+++ docs/conf.py
@@ -54,8 +54,8 @@ pygments_style = "sphinx"
 html_theme = "sphinx_rtd_theme"
 html_logo = "hostphot_logo.png"
 html_theme_options = {
-    'logo_only': False,
-    'display_version': True,
+    "logo_only": False,
+    "display_version": True,
 }
 html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 
