+++ src/hostphot/filters/config.txt
@@ -1,0 +1,7 @@
+survey	filters		zp		pixel_scale	mag_sys		mag_sys_conv
+DES	grizY		30		0.263		AB		0.0
+PS1	grizy		25		0.25		AB		0.0
+SDSS	ugriz		22.5		0.396		AB		0.0
+GALEX	FUV,NUV		18.82,30.08	1.5		AB		0.0
+2MASS	JHK		header		1.0		Vega		0.91,1.39,1.85
+WISE	W1,W2,W3,W4	header		1.375		Vega		2.699,3.339,5.174,6.620
