### 2022-07-20 temuller/hostphot#3 @krachyon
I'll mark this as resolved with the review done...
