### 2022-07-20 #4508 @Onoddil
Hi @temuller, @arfon,

I've just submitted my review issue over at https://github.com/temuller/hostphot/issues/4. It should hopefully be mostly relatively minor suggestions and some clarifications on my end, but let me know if there are any issues with the review.
