### 2022-07-20 #4508 @Onoddil
One issue I did have was that the paper seems to be out of sync with the markdown. The markdown preview on github is quite good so I believe my review to be unaffected by this, but just in case I'll re-review the paper if I can get the PDF version to update...
