### 2022-07-25 #4508 @arfon
@Onoddil - commands to @editorialbot have to be at the start of a comment.
