### 2022-07-25 1a1576d
changes following referee comments (issue #4)
