+++ paper/paper.bib
@@ -260,6 +260,49 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{matplotlib,
+  author={J. D. {Hunter}},
+  journal={Computing in Science Engineering}, 
+  title={Matplotlib: A 2D Graphics Environment}, 
+  year={2007},
+  volume={9},
+  number={3},
+  pages={90-95}
+}
+
+@Article{numpy,
+	 title         = {Array programming with {NumPy}},
+	 author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+		         van der Walt and Ralf Gommers and Pauli Virtanen and David
+		         Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+		         Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+		         and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+		         Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+		         R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+		         G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+		         Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+		         Travis E. Oliphant},
+	 year          = {2020},
+	 month         = sep,
+	 journal       = {Nature},
+	 volume        = {585},
+	 number        = {7825},
+	 pages         = {357--362},
+	 doi           = {10.1038/s41586-020-2649-2},
+	 publisher     = {Springer Science and Business Media {LLC}},
+	 url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+@inproceedings{pandas,
+  title={Data structures for statistical computing in python},
+  author={McKinney, Wes and others},
+  booktitle={Proceedings of the 9th Python in Science Conference},
+  volume={445},
+  pages={51--56},
+  year={2010},
+  organization={Austin, TX}
+}
+  
 @article{Perlmutter1999,
        author = {{Perlmutter}, S. and {Aldering}, G. and {Goldhaber}, G. and
          {Knop}, R.~A. and {Nugent}, P. and {Castro}, P.~G. and {Deustua}, S. and
@@ -340,6 +383,22 @@ archivePrefix = {arXiv},
   doi          = {10.5281/zenodo.5796924},
   url          = {https://doi.org/10.5281/zenodo.5796924}
 }
+
+@MISC{pyvo,
+       author = {{Graham}, Matthew and {Plante}, Ray and {Tody}, Doug and {Fitzpatrick}, Mike},
+        title = "{PyVO: Python access to the Virtual Observatory}",
+     keywords = {Software},
+ howpublished = {Astrophysics Source Code Library, record ascl:1402.004},
+         year = 2014,
+        month = feb,
+          eid = {ascl:1402.004},
+        pages = {ascl:1402.004},
+archivePrefix = {ascl},
+       eprint = {1402.004},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2014ascl.soft02004G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 
 @article{Riess1998,
        author = {{Riess}, Adam G. and {Filippenko}, Alexei V. and {Challis}, Peter and
