+++ docs/setup/installation.rst
@@ -26,3 +26,25 @@ To install the code from source, do the following:
 	cd hostphot
 	python -m pip install .
 
+
+Requirements
+########################
+
+HostPhot has the following requirements:
+
+```code
+numpy
+pandas
+matplotlib
+astropy
+reproject
+photutils
+astroquery
+extinction
+sfdmap
+pyvo
+sep
+ipywidgets (optional: for interactive aperture)
+ipympl (optional: for interactive aperture)
+pytest (optional: for testing the code)
+```
