### 2022-07-25 6c7c580
adding GALEX and updating error propagation
