+++ src/hostphot/cutouts.py
@@ -9,6 +9,7 @@ from astropy.table import Table
 from astropy import wcs, units as u
 from astropy.coordinates import SkyCoord
 
+import pyvo  # 2MASS
 from pyvo.dal import sia  # DES
 from astroquery.sdss import SDSS
 from astroquery.skyview import SkyView  # other surveys
@@ -20,11 +21,11 @@ from reproject.mosaicking import find_optimal_celestial_wcs
 from hostphot._constants import __workdir__
 from hostphot.utils import (
     get_survey_filters,
-    survey_pixel_scale,
     clean_dir,
     check_work_dir,
     check_survey_validity,
     check_filters_validity,
+    survey_pixel_scale,
 )
 from hostphot.image_cleaning import trim_images
 
@@ -56,7 +57,7 @@ def query_ps1(ra, dec, size=3, filters=None):
     dec: float
         Declination in degrees.
     size: float or ~astropy.units.Quantity, default ``3``
-        Image size. If a float is given the units are assumed to be in arcmin.
+        Image size. If a float is given, the units are assumed to be arcmin.
     filters: str, default ``None``
         Filters to use. If ``None``, uses ``grizy``.
 
@@ -98,7 +99,7 @@ def get_PS1_urls(ra, dec, size=3, filters=None):
     dec: float
         Declination in degrees.
     size: float or ~astropy.units.Quantity, default ``3``
-        Image size. If a float is given the units are assumed to be in arcmin.
+        Image size. If a float is given, the units are assumed to be arcmin.
     filters: str, default ``None``
         Filters to use. If ``None``, uses ``grizy``.
 
@@ -148,7 +149,7 @@ def get_PS1_images(ra, dec, size=3, filters=None):
     dec: float
         Declination in degrees.
     size: float or ~astropy.units.Quantity, default ``3``
-        Image size. If a float is given the units are assumed to be in arcmin.
+        Image size. If a float is given, the units are assumed to be arcmin.
     filters: str, default ``None``
         Filters to use. If ``None``, uses ``grizy``.
 
@@ -251,7 +252,7 @@ def get_DES_images(ra, dec, size=3, filters=None):
     dec: float
         Declination in degrees.
     size: float or ~astropy.units.Quantity, default ``3``
-        Image size. If a float is given the units are assumed to be in arcmin.
+        Image size. If a float is given, the units are assumed to be arcmin.
     filters: str, default ``None``
         Filters to use. If ``None``, uses ``grizY``.
 
@@ -294,7 +295,7 @@ def get_DES_images(ra, dec, size=3, filters=None):
 # ----------------------------------------
 def get_SDSS_images(ra, dec, size=3, filters=None):
     """Downloads a set of SDSS fits images for a given set
-    of coordinates and filters using astroquery.
+    of coordinates and filters using SkyView.
 
     Parameters
     ----------
@@ -302,8 +303,8 @@ def get_SDSS_images(ra, dec, size=3, filters=None):
         Right ascension in degrees.
     dec: str or float
         Declination in degrees.
-    size: float or ~astropy.units.Quantity, default ``3``
-        Image size. If a float is given the units are assumed to be in arcmin.
+    size: int, default ``3``
+        Image size in arcmin.
     filters: str, default ``None``
         Filters to use. If ``None``, uses ``ugriz``.
 
@@ -313,37 +314,47 @@ def get_SDSS_images(ra, dec, size=3, filters=None):
         List with fits images for the given filters.
         `None` is returned if no image is found.
     """
-    check_filters_validity(filters, "SDSS")
+    survey = "SDSS"
+    check_filters_validity(filters, survey)
     if filters is None:
[…]
