+++ src/hostphot/cutouts.py
@@ -159,9 +159,9 @@ def get_PS1_images(ra, dec, size=3, filters=None):
         List of fits images.
     """
     survey = 'PS1'
-    check_filters_validity(filters, survey)
     if filters is None:
         filters = get_survey_filters(survey)
+    check_filters_validity(filters, survey)
 
     fits_url = get_PS1_urls(ra, dec, size, filters)
 
@@ -197,9 +197,9 @@ def get_DES_urls(ra, dec, fov, filters="grizY"):
     url_w_list: list
         List of URLs with DES images weights.
     """
-    check_filters_validity(filters, "DES")
     if filters is None:
         filters = get_survey_filters("DES")
+    check_filters_validity(filters, "DES")
 
     des_access_url = "https://datalab.noirlab.edu/sia/des_dr1"
     svc = sia.SIAService(des_access_url)
@@ -262,9 +262,9 @@ def get_DES_images(ra, dec, size=3, filters=None):
         List of fits images.
     """
     survey = 'DES'
-    check_filters_validity(filters, survey)
     if filters is None:
         filters = get_survey_filters(survey)
+    check_filters_validity(filters, survey)
 
     if isinstance(size, (float, int)):
         fov = (size * u.arcmin).to(u.degree).value
@@ -315,9 +315,9 @@ def get_SDSS_images(ra, dec, size=3, filters=None):
         `None` is returned if no image is found.
     """
     survey = "SDSS"
-    check_filters_validity(filters, survey)
     if filters is None:
         filters = get_survey_filters(survey)
+    check_filters_validity(filters, survey)
 
     # SkyView calls the filters in a different way
     filters_dict = {'u': 'SDSSu',
@@ -391,16 +391,11 @@ def get_exptime(used_image, obs_table, filt):
     """
     galex_df = obs_table.to_pandas()
     galex_df = galex_df[galex_df.filters == filt]
-    urls = galex_df.dataURL.values
-    t_exptime = galex_df.t_exptime.values
 
-    texp = 0
-    for url, t in zip(urls, t_exptime):
-        if used_image in url:
-            texp = t
-            break
+    target_df = galex_df[galex_df.target_name==used_image]
+    texp = target_df.t_exptime.values[0]
 
-    if texp == 0:
+    if texp is None:
         raise valueError('Galex image not found.')
 
     return texp
@@ -427,9 +422,9 @@ def get_GALEX_images(ra, dec, size=3, filters=None):
         `None` is returned if no image is found.
     """
     survey = "GALEX"
-    check_filters_validity(filters, survey)
     if filters is None:
         filters = get_survey_filters(survey)
+    check_filters_validity(filters, survey)
 
     # SkyView calls the filters in a different way
     filters_dict = {'NUV': 'GALEX Near UV',
@@ -444,25 +439,26 @@ def get_GALEX_images(ra, dec, size=3, filters=None):
     pixel_scale = survey_pixel_scale(survey)
     size_pixels = int(size_arcsec.value / pixel_scale)
 
-    fits_files = []
     with warnings.catch_warnings():
         warnings.simplefilter('ignore', AstropyWarning)
         coords = SkyCoord(ra=ra, dec=dec, unit=(u.degree, u.degree), frame='icrs')
         obs_table = Observations.query_criteria(coordinates=coords, radius=size_arcsec,
                                                 obs_collection=[survey])
-        for filt, skyview_filter in zip(filters, skyview_filters):
-            fits_file = SkyView.get_images(position=coords,
+
+        skyview_fits = SkyView.get_images(position=coords,
                                            coordinates='icrs', pixels=str(size_pixels),
-                                           survey=skyview_filter, width=size_arcsec,
[…]
