+++ src/hostphot/coadd.py
@@ -9,18 +9,6 @@ from hostphot._constants import __workdir__
 import warnings
 from astropy.utils.exceptions import AstropyWarning
 
-# ----------------------------------------
-def _choose_workdir(workdir):
-    """Updates the work directory.
-
-    Parameters
-    ----------
-    workdir: str
-        Path to the work directory.
-    """
-    global __workdir__
-    __workdir__ = workdir
-
 
 # ----------------------------------------
 
