+++ tests/test_preprocessing.py
@@ -8,29 +8,30 @@ class TestHostPhot(unittest.TestCase):
     def test_preprocessing(self):
         coadd_filters = "riz"
         survey = "PS1"
-        name = "SN2004eo"
-        host_ra, host_dec = (
-            308.2092,
-            9.92755,
-        )  # coods of host galaxy of SN2004eo
-        download_images(name, host_ra, host_dec, survey=survey)
+        sn_name = "2002fk"
+        host_ra = 50.527333
+        host_dec = -15.400056
+
+        download_images(sn_name, host_ra, host_dec, survey=survey)
 
         # coadd
-        coadd_images(name, coadd_filters, survey)
+        coadd_images(sn_name, coadd_filters, survey)
 
         # masking
         coadd_mask_params = create_mask(
-            name,
+            sn_name,
             host_ra,
             host_dec,
             filt=coadd_filters,
             survey=survey,
             extract_params=True,
+            crossmatch=True,
         )
 
+        # apply mask on single-filter images
         for filt in "grizy":
             create_mask(
-                name,
+                sn_name,
                 host_ra,
                 host_dec,
                 filt,
