+++ src/hostphot/global_photometry.py
@@ -43,18 +43,6 @@ from astropy.utils.exceptions import AstropyWarning
 
 sep.set_sub_object_limit(1e4)
 
-# ----------------------------------------
-def _choose_workdir(workdir):
-    """Updates the work directory.
-
-    Parameters
-    ----------
-    workdir: str
-        Path to the work directory.
-    """
-    global __workdir__
-    __workdir__ = workdir
-
 
 # -------------------------------
 def kron_flux(data, err, gain, objects, kronrad, scale):
