+++ src/hostphot/rgb_images.py
@@ -12,94 +12,6 @@ from astropy.table import Table
 from hostphot.utils import survey_pixel_scale
 
 
-def sky_median_sig_clip(input_arr, sig_fract, percent_fract, max_iter=100):
-    """Estimates median sky value for a given number of iterations.
-
-    Parameters
-    ----------
-    input_arr: numpy array
-            Image data array
-    sig_fract: float
-            Fraction of sigma clipping.
-    percent_fract: float
-    Convergence fraction.
-    max_iter: int, default ``100``
-    Maximum number of iterations.
-
-    Returns
-    -------
-    (new_sky, iteration): tuple
-            Median sky value and number of iteration.
-    """
-    work_arr = np.ravel(input_arr)
-    old_sky = np.median(work_arr)
-    sig = work_arr.std()
-    upper_limit = old_sky + sig_fract * sig
-    lower_limit = old_sky - sig_fract * sig
-    indices = np.where((work_arr < upper_limit) & (work_arr > lower_limit))
-    work_arr = work_arr[indices]
-    new_sky = np.median(work_arr)
-    iteration = 0
-
-    while ((math.fabs(old_sky - new_sky) / new_sky) > percent_fract) and (
-        iteration < max_iter
-    ):
-        iteration += 1
-        old_sky = new_sky
-        sig = work_arr.std()
-        upper_limit = old_sky + sig_fract * sig
-        lower_limit = old_sky - sig_fract * sig
-        indices = np.where((work_arr < upper_limit) & (work_arr > lower_limit))
-        work_arr = work_arr[indices]
-        new_sky = np.median(work_arr)
-
-    return (new_sky, iteration)
-
-
-def sky_mean_sig_clip(input_arr, sig_fract, percent_fract, max_iter=100):
-    """Estimates mean sky value for a given number of iterations.
-
-    Parameters
-    ----------
-    input_arr: numpy array
-            Image data array
-    sig_fract: float
-            Fraction of sigma clipping.
-    percent_fract: float
-    Convergence fraction.
-    max_iter: int, default ``100``
-    Maximum number of iterations.
-
-    Returns
-    -------
-    (new_sky, iteration): tuple
-            Mean sky value and number of iteration.
-    """
-    work_arr = np.ravel(input_arr)
-    old_sky = np.mean(work_arr)
-    sig = work_arr.std()
-    upper_limit = old_sky + sig_fract * sig
-    lower_limit = old_sky - sig_fract * sig
-    indices = np.where((work_arr < upper_limit) & (work_arr > lower_limit))
-    work_arr = work_arr[indices]
-    new_sky = np.mean(work_arr)
-    iteration = 0
-
-    while ((math.fabs(old_sky - new_sky) / new_sky) > percent_fract) and (
-        iteration < max_iter
-    ):
-        iteration += 1
-        old_sky = new_sky
-        sig = work_arr.std()
-        upper_limit = old_sky + sig_fract * sig
-        lower_limit = old_sky - sig_fract * sig
-        indices = np.where((work_arr < upper_limit) & (work_arr > lower_limit))
-        work_arr = work_arr[indices]
-        new_sky = np.mean(work_arr)
-
-    return (new_sky, iteration)
-
-
 def linear(inputArray, scale_min=None, scale_max=None):
     """Performs linear scaling of the input numpy array.
 
