### 2022-07-28 #4508 @temuller
@Onoddil I have made several changes, following your suggestions from [temuller/hostphot#4](https://github.com/temuller/hostphot/issues/4), which I have merged with the main branch. Thanks for you detailed report. Let me know if you think there is still some things missing.
