### 2022-07-28 temuller/hostphot#4 @Onoddil
To avoid quote replying to your quote reply and it being a bit of a mess I'll just number my replies to your replies.

1. Great, thanks!
2. Perfect
3. That's fine, I just had to check as per the demands of the review list. If there's no direct comparison then no comparison can be made.
4. Thanks
5. It would be good to expand the tests as time goes on and new features are added, but I agree that the nature of quite a lot of the functionality limits the verification step somewhat (how do you check for "I downloaded an image?" in a way that verifies the image is exactly the same, or how do you check that an interactive GUI works in an automated script?), so it's good that you've added bits where possible.
6. Coverage is now at 80% as you mentioned, which is much more complete than previously. With such a small codebase small bits of untestable code will have a large impact on coverage, so I'm not too sure how you'd get much higher without having to re-write a lot of the code or spending a lot of time. Again, it would be good to make sure coverage remains high as time goes on, but that's outside of the scope of the review!
7. That's a good way to handle contact details, and thanks for adding the redirect.
8. Thank you for adding examples to the docstrings.
9. Just leaving the directory as a fixed folder name is probably fine to be honest, and avoids all of this complication. But yes, the idea would be to not have a `hostphot.cutouts.__workdir__` (or have one per individual `.py` file anymore, as you have now), you'd always use the one in `hostphot._constants` instead. But the way you have it set up now does work, so I was just being nitpicky in trying to tidy up the code; either keeping it as it is, seeing if you like my suggestion, or leaving the directory as a fixed value would all be fine, none of that is relevant for my review at the moment so I'll just leave it as-is.
10. The new wording is much clearer, thanks for clarifying that.

This answers all of my questions from the review so I'll close this issue!
