### 2022-07-28 #4508 @Onoddil
Thanks for the thorough responses @temuller, everything looks good now following the review in https://github.com/temuller/hostphot/issues/4 so I'm happy to recommend hostphot for publication!
cc @arfon
