+++ src/hostphot/utils.py
@@ -150,6 +150,9 @@ def get_image_gain(header, survey):
         gain = header["GAIN"]
     elif survey == "SDSS":
         gain = 1.0
+    elif survey == "2MASS":
+        # https://iopscience.iop.org/article/10.1086/498708/pdf
+        gain = 8.0
     else:
         gain = 1.0
 
@@ -183,6 +186,10 @@ def get_image_readnoise(header, survey):
         readnoise = 7.0  # electrons per pixel
     elif survey == "SDSS":
         readnoise = 0.0
+    elif survey=="2MASS":
+        # https://iopscience.iop.org/article/10.1086/498708/pdf
+        # 6 combined images
+        readnoise = 4.5*np.sqrt(6)  # not used
     else:
         readnoise = 0.0
 
@@ -206,12 +213,22 @@ def get_image_exptime(header, survey):
     check_survey_validity(survey)
     if survey in ["PS1", "DES", "GALEX"]:
         exptime = float(header["EXPTIME"])
+    elif survey=="WISE":
+        # see: https://wise2.ipac.caltech.edu/docs/release/allsky/
+        # and https://wise2.ipac.caltech.edu/docs/release/allsky/expsup/sec1_1.html
+        if header["BAND"] in [1, 2]:
+            exptime = 7.7
+        elif header["BAND"] in [3, 4]:
+            exptime = 8.8
+    elif survey == "2MASS":
+        # https://iopscience.iop.org/article/10.1086/498708/pdf
+        exptime = 7.8
     else:
-        exptime = 0.0
+        exptime = 1.0
 
     return exptime
 
-def uncertainty_calc(flux, survey, filt=None, ap_area=0.0, readnoise=0.0, gain=1.0, exptime=0.0):
+def uncertainty_calc(flux, flux_err, survey, filt=None, ap_area=0.0, readnoise=0.0, gain=1.0, exptime=0.0, bkg_rms=0.0):
     """Calculates the uncertainty propagation.
 
     Parameters
@@ -226,10 +243,12 @@ def uncertainty_calc(flux, survey, filt=None, ap_area=0.0, readnoise=0.0, gain=1
         Aperture area.
     readnoise: float, default ``0.0``
         Image readnoise.
-    gain: str, default ``1.0``
+    gain: float, default ``1.0``
         Image gain.
-    exptime: str, default ``0.0``
+    exptime: float, default ``0.0``
         Image exposure time.
+    bkg_rms: float, default ``0.0``
+        Background noise.
 
     Returns
     -------
@@ -247,7 +266,7 @@ def uncertainty_calc(flux, survey, filt=None, ap_area=0.0, readnoise=0.0, gain=1
         mag_err = np.sqrt(mag_err ** 2 + extra_err ** 2)
 
     elif survey=="GALEX":
-        CPS = flux.copy()
+        CPS = flux
         if filt == 'FUV':
             uv_err = -2.5 * (
                         np.log10(CPS) - np.log10(CPS + (CPS * exptime + (0.050 * CPS * exptime) ** 2) ** 0.5 / exptime))
@@ -255,6 +274,39 @@ def uncertainty_calc(flux, survey, filt=None, ap_area=0.0, readnoise=0.0, gain=1
             uv_err = -2.5 * (
                         np.log10(CPS) - np.log10(CPS + (CPS * exptime + (0.027 * CPS * exptime) ** 2) ** 0.5 / exptime))
         mag_err = np.sqrt(mag_err ** 2 + uv_err ** 2)
+
+    elif survey=="2MASS":
[…]
