+++ src/hostphot/utils.py
@@ -12,8 +12,9 @@ import warnings
 from astropy.utils.exceptions import AstropyWarning
 
 import hostphot
+
 hostphot_path = hostphot.__path__[0]
-config_file = os.path.join(hostphot_path, 'filters', 'config.txt')
+config_file = os.path.join(hostphot_path, "filters", "config.txt")
 config_df = pd.read_csv(config_file, delim_whitespace=True)
 
 
@@ -59,9 +60,7 @@ def check_survey_validity(survey):
     """
     global config_df
     surveys = list(config_df.survey)
-    assert survey in surveys, (
-        f"survey '{survey}' not" f" in {surveys}"
-    )
+    assert survey in surveys, f"survey '{survey}' not" f" in {surveys}"
 
 
 def get_survey_filters(survey):
@@ -83,8 +82,8 @@ def get_survey_filters(survey):
     survey_df = config_df[config_df.survey == survey]
     filters = survey_df.filters.values[0]
 
-    if ',' in filters:
-        filters = filters.split(',')
+    if "," in filters:
+        filters = filters.split(",")
 
     return filters
 
@@ -113,11 +112,11 @@ def survey_zp(survey):
     survey_df = config_df[config_df.survey == survey]
     zps = survey_df.zp.values[0]
 
-    if zps=='header':
+    if zps == "header":
         return zps
 
-    if ',' in zps:
-        zps = zps.split(',')
+    if "," in zps:
+        zps = zps.split(",")
         zp_dict = {filt: float(zp) for filt, zp in zip(filters, zps)}
     else:
         zp_dict = {filt: float(zps) for filt in filters}
@@ -186,14 +185,15 @@ def get_image_readnoise(header, survey):
         readnoise = 7.0  # electrons per pixel
     elif survey == "SDSS":
         readnoise = 0.0
-    elif survey=="2MASS":
+    elif survey == "2MASS":
         # https://iopscience.iop.org/article/10.1086/498708/pdf
         # 6 combined images
-        readnoise = 4.5*np.sqrt(6)  # not used
+        readnoise = 4.5 * np.sqrt(6)  # not used
     else:
         readnoise = 0.0
 
     return readnoise
+
 
 def get_image_exptime(header, survey):
     """Returns the exposure time from an image's header.
@@ -213,7 +213,7 @@ def get_image_exptime(header, survey):
     check_survey_validity(survey)
     if survey in ["PS1", "DES", "GALEX"]:
         exptime = float(header["EXPTIME"])
-    elif survey=="WISE":
+    elif survey == "WISE":
         # see: https://wise2.ipac.caltech.edu/docs/release/allsky/
         # and https://wise2.ipac.caltech.edu/docs/release/allsky/expsup/sec1_1.html
         if header["BAND"] in [1, 2]:
@@ -228,7 +228,18 @@ def get_image_exptime(header, survey):
 
     return exptime
 
-def uncertainty_calc(flux, flux_err, survey, filt=None, ap_area=0.0, readnoise=0.0, gain=1.0, exptime=0.0, bkg_rms=0.0):
+
+def uncertainty_calc(
+    flux,
+    flux_err,
+    survey,
+    filt=None,
+    ap_area=0.0,
+    readnoise=0.0,
+    gain=1.0,
+    exptime=0.0,
+    bkg_rms=0.0,
+):
     """Calculates the uncertainty propagation.
 
     Parameters
@@ -259,54 +270,80 @@ def uncertainty_calc(flux, flux_err, survey, filt=None, ap_area=0.0, readnoise=0
     if survey in ["PS1", "DES", "SDSS"]:
[…]
