+++ src/hostphot/cutouts.py
@@ -1,5 +1,7 @@
 import os
 import copy
+import tarfile
+import requests
 import numpy as np
 import pandas as pd
 import multiprocessing as mp
@@ -151,12 +153,12 @@ def get_PS1_images(ra, dec, size=3, filters=None):
 
     fits_url = get_PS1_urls(ra, dec, size, filters)
 
-    fits_files = []
+    hdu_list = []
     for url, filt in zip(fits_url, filters):
-        fits_file = fits.open(url)
-        fits_files.append(fits_file)
+        hdu = fits.open(url)
+        hdu_list.append(hdu)
 
-    return fits_files
+    return hdu_list
 
 
 # DES
@@ -244,7 +246,7 @@ def get_DES_images(ra, dec, size=3, filters=None):
 
     Returns
     -------
-    fits_files: list
+    hdu_list: list
         List of fits images.
     """
     survey = "DES"
@@ -262,19 +264,19 @@ def get_DES_images(ra, dec, size=3, filters=None):
     if url_list is None:
         return None
 
-    fits_files = []
+    hdu_list = []
     for url, url_w in zip(url_list, url_w_list):
         # combine image+weights on a single fits file
-        image_fits = fits.open(url)
-        weight_fits = fits.open(url_w)
-        hdu = fits.PrimaryHDU(image_fits[0].data, header=image_fits[0].header)
+        image_hdu = fits.open(url)
+        weight_hdu = fits.open(url_w)
+        hdu = fits.PrimaryHDU(image_hdu[0].data, header=image_hdu[0].header)
         hdu_err = fits.ImageHDU(
-            weight_fits[0].data, header=weight_fits[0].header
+            weight_hdu[0].data, header=weight_hdu[0].header
         )
-        hdu_list = fits.HDUList([hdu, hdu_err])
-        fits_files.append(hdu_list)
+        hdu_sublist = fits.HDUList([hdu, hdu_err])
+        hdu_list.append(hdu_sublist)
 
-    return fits_files
+    return hdu_list
 
 
 # SDSS
@@ -296,7 +298,7 @@ def get_SDSS_images(ra, dec, size=3, filters=None):
 
     Return
     ------
-    fits_files: list
+    hdu_list: list
         List with fits images for the given filters.
         `None` is returned if no image is found.
     """
@@ -323,15 +325,15 @@ def get_SDSS_images(ra, dec, size=3, filters=None):
     pixel_scale = survey_pixel_scale(survey)
     size_pixels = int(size_arcsec.value / pixel_scale)
 
-    fits_files = []
     with warnings.catch_warnings():
         warnings.simplefilter("ignore", AstropyWarning)
         coords = SkyCoord(
             ra=ra, dec=dec, unit=(u.degree, u.degree), frame="icrs"
         )
 
+        hdu_list = []
         for filt, skyview_filter in zip(filters, skyview_filters):
-            fits_file = SkyView.get_images(
+            hdu = SkyView.get_images(
                 position=coords,
                 coordinates="icrs",
                 pixels=str(size_pixels),
@@ -340,9 +342,9 @@ def get_SDSS_images(ra, dec, size=3, filters=None):
                 height=size_arcsec,
             )
 
-            fits_files.append(fits_file[0])
+            hdu_list.append(hdu[0])
 
-    return fits_files
+    return hdu_list
 
 
 # GALEX
@@ -410,7 +412,7 @@ def get_GALEX_images(ra, dec, size=3, filters=None):
 
     Return
     ------
-    fits_files: list
+    hdu_list: list
         List with fits images for the given filters.
         `None` is returned if no image is found.
     """
[…]
