### 2022-08-10 900d233
adding description of how HostPhot works to the docs
