### 2022-08-15 #4508 @arfon
@temuller – could you merge this PR please? https://github.com/temuller/hostphot/pull/5
